Made it installable
With pylint2sarif and pylint2cso commands (scripts).

I've kept the modules top-level, i.e., made each of them its own
"package" (in the pip sense, not in the sense of being directories
instead of .py files). But I've moved them into a src/ directory so
it's easier to avoid running them when they're *not* installed, if
one wishes to avoid this.

I'm using poetry as a dependency manager, since that nicely
supports using the pyproject.toml format and tracking separate
frozen package versions of direct and indirect dependencies (in
poetry.lock).

I've made pylint itself just a dev dependency, because users may
prefer to use pylint as installed systemwide or in their own
project's environment. python-jsonschema-objects is of course a
regular ("main") dependency.

Both "poetry install" in the source tree, and building and
installing wheels, seem to be working okay.

Currently I have a symlink in src to the sarif-schema.json file
that remains in the parent directory, which is a hackish workaround
so "poetry install" is sufficient to get a fully usable
installation. I think there is some better way to achieve this;
that's what the "TODO" comment in pyproject.toml is about. Building
and installing wheels works fine regardless of whether this symlink
is present.