Update dependencies
This neither fixes nor exacerbates #19. Because I am still giving
the 3.7.1 as the earliest version to support, poetry installs an
early enough version of jsonschema to avoid #19, even though that
bug is conceptually unrelated to what version of Python is used.
Unfortunately, this only applies to the "poetry install" usage,
which is not a significant use case other than for development.
The problem in #19 will usually (perhaps almost always) happen in
actual use.