Hold jsonschema at 4.17.* to work around #19
This also updates the readme accordingly, to link #19 and recommend
installing via pipx instead of as a dev dependency of a project
that might otherwise depend on jsonschema and benefit from using a
later version.

It also adds a link to PyPI. This will look a little weird as long
as the main (repo root) readme and PyPI readme are the same file,
but I think it's better than omitting it, and I don't want to add a
badge for it instead until I've thought through what other badges
(if any) ought to be shown.